% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assignment_meta_data.R
\name{Assignment_meta_data}
\alias{Assignment_meta_data}
\title{Bulk Assignments to meta data}
\usage{
Assignment_meta_data(
  assignment_df,
  clusters,
  output_csv = TRUE,
  output_file = "SoupLadle_Assignments.csv"
)
}
\arguments{
\item{assignment_df}{Assignemnt data.frame from the SNP_based_assignment}

\item{clusters}{Demultiplexing file of the scRNA-seq data (e.g. cluster.tsv from SouporCell)}

\item{output_csv}{Logical of wheter an output file will be generated}

\item{output_file}{Path for the output csv file that can be used as meta data in any scRNA-seq processing object}
}
\value{
meta data for Seurat
}
\description{
Bulk Assignments to meta data
}
