% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binarize_genotype_values.R
\name{Binarize_genotype_values}
\alias{Binarize_genotype_values}
\title{Binarization of the SNP profile}
\usage{
Binarize_genotype_values(matrix)
}
\arguments{
\item{matrix}{SNP matrix that will be binarized}
}
\value{
binarized SNP matrix
}
\description{
Binarization of the SNP profile
}
