% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_extract_vcf.R
\name{read_extract_vcf}
\alias{read_extract_vcf}
\title{Read vcf files}
\usage{
read_extract_vcf(vcf_file, gt = NULL, common_SNP_filter = TRUE)
}
\arguments{
\item{vcf_file}{Path to vcf file to load (e.g. bulkRNA-seq or scRNA-seq variants)}

\item{gt}{element to extract from vcf genotype data. Common options include 'DP', 'GT' and 'GQ'.}

\item{common_SNP_filter}{Filter out all SNPs that have a common genotype in all samples or clusters}
}
\value{
A tibble of the loaded vcf file
}
\description{
Read vcf files
}
