% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_based_assignment.R
\name{SNP_based_assignment}
\alias{SNP_based_assignment}
\title{Dissimilarity Matrix of binarized SNP matrix}
\usage{
SNP_based_assignment(matrix1, matrix2)
}
\arguments{
\item{matrix1}{Binarized SNP matrix from demultiplexed single-cell RNA-seq experiments}

\item{matrix2}{Binarized SNP matrix from called SNPs of bulkRNA}
}
\value{
Dissimilarity Matrix
}
\description{
Dissimilarity Matrix of binarized SNP matrix
}
